+++ state/executor.go
@@ -198,7 +198,7 @@ func (e *BlockExecutor) updateState(state types.State, block *types.Block, final
 				}
 			}
 			// Change results from this height but only applies to the next next height.
-			lastHeightValSetChanged = block.Height() + 1 + 1
+			lastHeightValSetChanged = int64(block.Height()) + 1 + 1
 		}
 
 		if len(nValSet.Validators) > 0 {
@@ -213,7 +213,7 @@ func (e *BlockExecutor) updateState(state types.State, block *types.Block, final
 		Version:         state.Version,
 		ChainID:         state.ChainID,
 		InitialHeight:   state.InitialHeight,
-		LastBlockHeight: block.Height(),
+		LastBlockHeight: int64(block.Height()),
 		LastBlockTime:   block.Time(),
 		LastBlockID: cmtypes.BlockID{
 			Hash: cmbytes.HexBytes(block.Hash()),
@@ -265,10 +265,10 @@ func (e *BlockExecutor) validate(state types.State, block *types.Block) error {
 		block.SignedHeader.Version.Block != state.Version.Consensus.Block {
 		return errors.New("block version mismatch")
 	}
-	if state.LastBlockHeight <= 0 && block.Height() != state.InitialHeight {
+	if state.LastBlockHeight <= 0 && int64(block.Height()) != state.InitialHeight {
 		return errors.New("initial block height mismatch")
 	}
-	if state.LastBlockHeight > 0 && block.Height() != state.LastBlockHeight+1 {
+	if state.LastBlockHeight > 0 && int64(block.Height()) != state.LastBlockHeight+1 {
 		return errors.New("block height mismatch")
 	}
 	if !bytes.Equal(block.SignedHeader.AppHash[:], state.AppHash[:]) {
@@ -377,7 +377,7 @@ func (e *BlockExecutor) publishEvents(resp *abci.ResponseFinalizeBlock, block *t
 		for _, ev := range abciBlock.Evidence.Evidence {
 			if err := e.eventBus.PublishEventNewEvidence(cmtypes.EventDataNewEvidence{
 				Evidence: ev,
-				Height:   block.SignedHeader.Header.Height(),
+				Height:   int64(block.SignedHeader.Header.Height()),
 			}); err != nil {
 				e.logger.Error("failed publishing new evidence", "err", err)
 			}
